+++ CCS/wisp-base/nvm/fram.h
@@ -20,8 +20,8 @@
 
 void FRAM_init(void);
 
-void FRAM_write_long_array(uint32_t *address , uint16_t numberOfLongs , uint32_t *writeData);
-void FRAM_write_int_array(uint16_t *address , uint16_t numberOfInts , uint16_t *writeData);
+void FRAM_write_long_array(unsigned long *address , uint16_t numberOfLongs , uint32_t *writeData);
+void FRAM_write_int_array(unsigned int *address , uint16_t numberOfInts , uint16_t *writeData);
 void FRAM_write_char_array(uint8_t *address , uint16_t numberOfChars , uint8_t *writeData);
 
 void FRAM_write_infoA_long(int addressOffset , uint16_t numberOfLongs , uint32_t *writeData);
